Remove EventDatabaseEventResolver and minor refactor of model
After more testing with attempts to normalize the schema, I
can't match the performance of simply loading the whole ProviderDetails,
so I'm removing EventDatabaseEventResolver.

This has the happy side effect of letting us use the same logic across
all providers, making it easy to compare functionality and performance.

Also removed the ValueName class. I'm not sure why I did it this way
before (might have had to do with marshalling between .NET and
Javascript in Electron), but it makes more sense for this to be a
KeyValuePair and just store it in a dictionary.