Move filtering and sorting into the Store
We were paying a performance penalty for trying to keep the Store
slim and letting the view do the sorting and filtering. Nearly
any state update, such as selecting an event, was causing the
view to re-filter and re-sort all the events, which became slow
with large event logs loaded.

With this change, all that work is done inside the reducers and is
exposed as part of the EventLogState. The reducers are able to
make better decisions on whether any large collections need to be
modified.