Reduce unnecessary sorting
When filtering, we only need to re-sort if the
order has been changed due to AsParallel, and
there's no performance gain when using AsParallel
for DateFilter and AdvancedFilter. Therefore, this
change reverts to the previous logic where those
two filter types are handled separately and do not
require a sort.