Allow loading logs with overlapping RecordId values
The @key attribute on the context menu created a situation
where logs with the same RecordId values cannot be loaded
simultaneously. This causes #252.

This fixes the problem by removing @key from the context menus.
It's not clear if this is needed for anything, but in some quick
testing it seems to work fine.

If this approach fails we will need to look at assigning all
loaded events some sort of unique ID that we can use as a key.